Addressed findings on utils package
