+++ com.teamscale.polarion.plugin/src/com/teamscale/polarion/plugin/utils/CastUtils.java
@@ -2,11 +2,9 @@ package com.teamscale.polarion.plugin.utils;
 
 import com.polarion.alm.projects.model.IUniqueObject;
 import com.polarion.alm.projects.model.IUser;
-import com.polarion.alm.tracker.model.IApprovalStruct;
 import com.polarion.alm.tracker.model.IAttachment;
 import com.polarion.alm.tracker.model.ICategory;
 import com.polarion.alm.tracker.model.IComment;
-import com.polarion.alm.tracker.model.IHyperlinkStruct;
 import com.polarion.alm.tracker.model.ILinkRoleOpt;
 import com.polarion.alm.tracker.model.ILinkedWorkItemStruct;
 import com.polarion.alm.tracker.model.IModule;
@@ -20,7 +18,6 @@ import com.teamscale.polarion.plugin.model.LinkDirection;
 import com.teamscale.polarion.plugin.model.LinkedWorkItem;
 import com.teamscale.polarion.plugin.model.UpdateType;
 import com.teamscale.polarion.plugin.model.WorkItemForJson;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashMap;
@@ -30,8 +27,12 @@ import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class Utils {
+/**
+ * General utility class for helping work item history and prepping objects for json serialization.
+ */
+public class CastUtils {
 
+  /** Polarion field name for linked work items -- */
   public static final String LINKED_WORK_ITEMS_FIELD_NAME = "linkedWorkItems";
 
   /**
@@ -136,7 +137,8 @@ public class Utils {
       workItemForJson.setCategories(castCollectionToStrList(workItem.getCategories()));
     }
     if (workItem.getHyperlinks() != null && !workItem.getHyperlinks().isEmpty()) {
-      workItemForJson.setHyperLinks(castHyperlinksToStrList(workItem.getHyperlinks()));
+      workItemForJson.setHyperLinks(
+          CollectionsAndEnumsUtils.castHyperlinksToStrList(workItem.getHyperlinks()));
     }
     if (workItem.getComments() != null && !workItem.getComments().isEmpty()) {
       workItemForJson.setComments(
@@ -176,102 +178,6 @@ public class Utils {
     }
 
     return workItemForJson;
-  }
-
-  /**
-   * Takes a raw collection of Polarion hyperlinks and converts to a list of strings (link URIs).
-   */
-  public static List<String> castHyperlinksToStrList(Collection hyperlinks) {
-    List<String> result = new ArrayList<>();
-    if (isCollectionHyperlinkStructList(hyperlinks)) {
-      try {
-        List<IHyperlinkStruct> collection = (List<IHyperlinkStruct>) hyperlinks;
-        result =
-            collection.stream()
-                .map(elem -> ((IHyperlinkStruct) elem).getUri())
-                .collect(Collectors.toList());
-      } catch (ClassCastException ex) {
-        // casting should not be an issue since we're checking it on the if
-        return result;
-      }
-    }
-    return result;
-  }
-
-  /**
-   * Takes a raw collection of Polarion linkStructs and converts to a list of strings (the linked
-   * items IDs).
-   */
-  public static List<String> castLinkedWorkItemsToStrList(Collection linkedItems) {
-    List<String> result = new ArrayList<>();
-    if (isCollectionLinkedWorkItemStructList(linkedItems)) {
-      try {
-        List<ILinkedWorkItemStruct> collection = (List<ILinkedWorkItemStruct>) linkedItems;
-        result =
-            collection.stream()
-                .map(elem -> elem.getLinkedItem().getId())
-                .collect(Collectors.toList());
-      } catch (ClassCastException ex) {
-        // casting should not be an issue since we're checking it on the if
-      }
-    }
-    return result;
-  }
-
-  /**
-   * Takes a raw collection of Polarion approvalStructs and converts to a list of strings (the users
-   * IDs).
-   */
-  public static List<String> castApprovalsToStrList(Collection approvals) {
-    List<String> result = new ArrayList<>();
-    if (isCollectionApprovalStructList(approvals)) {
-      try {
-        List<IApprovalStruct> collection = (List<IApprovalStruct>) approvals;
-        result =
-            collection.stream()
-                .map(elem -> ((IApprovalStruct) elem).getUser().getId())
-                .collect(Collectors.toList());
-      } catch (ClassCastException ex) {
-        // casting should not be an issue since we're checking it on the if
-      }
-    }
-    return result;
-  }
-
-  /** This will return false if the list is empty, even if the list is of type IHyperlinkStruct */
-  public static boolean isCollectionHyperlinkStructList(Collection collection) {
-    if (collection instanceof List) {
-      List<?> list = (List<?>) collection;
-      if (!list.isEmpty() && list.get(0) instanceof IHyperlinkStruct) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /**
-   * This will return false if the list is empty, even if the list is of type
-   * ILinkedWorkedItemStruct
-   */
-  public static boolean isCollectionLinkedWorkItemStructList(Collection collection) {
-    if (collection instanceof List) {
-      List<?> list = (List<?>) collection;
-      if (!list.isEmpty() && list.get(0) instanceof ILinkedWorkItemStruct) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  /** This will return false if the list is empty, even if the list is of type IApprovalStruct */
-  public static boolean isCollectionApprovalStructList(Collection collection) {
-    if (collection instanceof List) {
-      List<?> list = (List<?>) collection;
-      if (!list.isEmpty() && list.get(0) instanceof IApprovalStruct) {
-        return true;
-      }
-    }
-    return false;
   }
 
   private static HashMap<String, Object> castCustomFields(
